exercice 2.5
