+++ part2/courseinfo/src/App.jsx
@@ -1,30 +1,4 @@
-const Header = ({course}) => <h1>{course}</h1>
-
-const Content = ({parts}) => (
-    <div>
-        {
-            parts.map(part => (
-                <Part key={part.id} part={part}/>
-            ))
-        }
-    </div>
-)
-
-const Part = ({part}) => (
-    <p>
-        {part.name} {part.exercises}
-    </p>
-)
-
-const Total = ({total}) => <p><b>Number of exercises {total}</b></p>
-
-const Course = ({course}) => (
-    <div>
-        <Header course={course.name}/>
-        <Content parts={course.parts}/>
-        <Total total={course.parts.reduce((sum, part) => sum + part.exercises, 0)}/>
-    </div>
-)
+import Course from "./components/Course.jsx";
 
 const App = () => {
     const courses = [
