+++ part2/courseinfo/src/components/Course.jsx
@@ -1,0 +1,29 @@
+const Course = ({course}) => (
+    <div>
+        <Header course={course.name}/>
+        <Content parts={course.parts}/>
+        <Total total={course.parts.reduce((sum, part) => sum + part.exercises, 0)}/>
+    </div>
+)
+
+const Header = ({course}) => <h1>{course}</h1>
+
+const Content = ({parts}) => (
+    <div>
+        {
+            parts.map(part => (
+                <Part key={part.id} part={part}/>
+            ))
+        }
+    </div>
+)
+
+const Part = ({part}) => (
+    <p>
+        {part.name} {part.exercises}
+    </p>
+)
+
+const Total = ({total}) => <p><b>Number of exercises {total}</b></p>
+
+export default Course
